chore: Enable caching in GitlabConfig and add cache management in GitlabClient
The GitlabConfig class has been updated to enable caching by adding the @EnableCaching annotation. This allows for caching of responses from Gitlab API, improving the performance of the proxy.

In the GitlabClient class, a cache manager has been injected and the getGroups() method has been annotated with @Cacheable("groupsCache"). This caches the response from Gitlab API, reducing the number of requests made and improving performance.