refactor: Modify getGroups() method in GroupsService to support cache refresh
The getGroups() method in the GroupsService class has been modified to support cache refresh. A new boolean parameter "refresh" has been added to the method signature, allowing the caller to specify whether to force a refresh of the groups data. This change improves the flexibility and usability of the service.